Fix crash
Whenever the while loop is entered it will remain inside infinitely.  This usually manifests in an out of memory error and crash.
The problem is that timeStamp never decreases but duration always does, so once timeStamp is greater than duration it always will be.  The intention here I'm certain is to decrease timestamp by duration.  This results in the correct animation behavior.